@@ -18,14 +18,13 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.14.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_hooks: ^0.15.0
-  hooks_riverpod: ^0.12.4
-  state_notifier: ^0.6.0
+  hooks_riverpod: ^1.0.0-dev.7
+  state_notifier: ^0.7.1
 
 
   # The following adds the Cupertino Icons font to your application.
